requirements.txt
torch>=2.2
hydra-core>=1.3
omegaconf>=2.3
aim>=3.19
mlflow>=2.14
tqdm>=4.66
matplotlib>=3.8
numpy>=1.26

